Checking in three test functions - to validate potential code issues

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware2018.java
@@ -192,7 +192,7 @@ public class hardware2018 {
         DriveTimed(direction, timeInMilliseconds, 1);
     }
 
-    public void DriveTimed(DriveDirection direction, int timeInMilliseconds, int Power) {
+    public void DriveTimed(DriveDirection direction, int timeInMilliseconds, double Power) {
 
         ElapsedTime runtime = new ElapsedTime();
         runtime.reset();
